requirements.txt
google-generativeai
python-dotenv==1.0.0
pillow
